# swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"PostResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"imageUrl": {
						"type": "string"
					},
					"caption": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"hasPinned": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"imageUrl",
					"caption",
					"createdAt",
					"userId",
					"username",
					"avatarUrl",
					"hasPinned"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PinsInfo": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"postId": {
						"type": "number",
						"format": "double"
					},
					"caption": {
						"type": "string",
						"nullable": true
					},
					"imageUrl": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"postId",
					"caption",
					"imageUrl",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserProfileResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"bio": {
						"type": "string",
						"nullable": true
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"followers": {
						"type": "number",
						"format": "double"
					},
					"following": {
						"type": "number",
						"format": "double"
					},
					"pins": {
						"items": {
							"$ref": "#/components/schemas/PinsInfo"
						},
						"type": "array"
					},
					"hasFollowed": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"username",
					"bio",
					"avatarUrl",
					"createdAt",
					"followers",
					"following",
					"pins",
					"hasFollowed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"passwordHash": {
						"type": "string"
					},
					"bio": {
						"type": "string",
						"nullable": true
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"email",
					"passwordHash",
					"bio",
					"avatarUrl",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Pinterest Simplified API",
		"version": "1.0.0",
		"description": "API docs for simplified Pinterest backend",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Your Name"
		}
	},
	"paths": {
...
		"/users/{userIdToFollow}/follow": {
			"post": {
				"operationId": "FollowUser",
				"responses": {
					"200": {
						"description": "Followed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userIdToFollow",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userIdToUnfollow}/unfollow": {
			"delete": {
				"operationId": "UnfollowUser",
				"responses": {
					"200": {
						"description": "Unfollowed",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userIdToUnfollow",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/profile": {
			"get": {
				"operationId": "GetUserProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserProfileResponse"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/pins": {
			"get": {
				"operationId": "GetPinners",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pins": {
											"items": {
												"$ref": "#/components/schemas/PinsInfo"
											},
											"type": "array"
										}
									},
									"required": [
										"pins"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/posts": {
			"get": {
				"operationId": "GetPostsByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PostResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/followers": {
			"get": {
				"operationId": "GetFollowers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"followers": {
											"items": {
												"$ref": "#/components/schemas/User"
											},
											"type": "array"
										}
									},
									"required": [
										"followers"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/following": {
			"get": {
				"operationId": "GetFollowing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"following": {
											"items": {
												"$ref": "#/components/schemas/User"
											},
											"type": "array"
										}
									},
									"required": [
										"following"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users & Follows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}